refactor: 숫자 야구 게임 helper 메서드 분리
