+++ docs/README.md
@@ -57,3 +57,4 @@
 - [x] 메서드 수정 (값 반환 -> 프로퍼티 업데이트)
 - [x] 숫자 야구 컴퓨터 분리
 - [x] 숫자 야구 유저 분리
+- [x] 게임 helper 메서드 분리
